+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Release Date: <insert date of release>
 - Added inline “+” buttons to enable quick addition of rows and columns across test steps, test data sheets, settings, and configurations
 - Implemented alphabetical sorting for test scenarios and test cases within the Test Lab
 - Implemented closing of popups on Esc keypress
+- Added Update (Rename) and Deleting for Test Case Tags
 
 #### Changed